+++ source/search_engine_forum_INEX.h
@@ -13,7 +13,7 @@ public:
 	ANT_search_engine_forum_INEX(char *filename, char *participant_id, char *run_id, char *task);
 	virtual ~ANT_search_engine_forum_INEX();
 
-	void write(long topic_id, char **docids, long hits);
+	void write(long topic_id, char **docids, long long hits);
 } ;
